move conversion funcs and effects to common location
