+++ ir_code_discovery_and_export/bit_flipir.py
@@ -2,7 +2,7 @@ import PySimpleGUI as sg
 import serial
 import clipboard
 import time
-from pixmob_conversion_funcs import to_arduino_string
+from shared.pixmob_conversion_funcs import to_arduino_string
 
 # BitFlipIR
 # This is a quick-and-dirty program requested by @Sean1983 to give a user a quick UI with some "bit"
