+++ shared/effect_definitions.py
